backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
python-dotenv==1.0.1
langchain==0.3.0
langchain-openai==0.2.0
solana==0.34.0
solders==0.21.0
anchorpy==0.20.1
pydantic==2.9.0